// src/app/demo/api/cehomologada.models.ts
export interface HomologadaCE {
    index_homenlaces: number;
    cod_gpo_ocup: string;
    desc_gpo_ocup: string;
    cod_area: string;
    desc_area: string;
    cod_actividad: string;
    desc_actividad: string;
    cod_subactividad: string;
    desc_subactividad: number;
    cod_servicio: string;
    desc_servicio: string;
    cod_especialidad: string;
    especialidad: number;
    cod_subespecialidad: string;
    subespecialidad: string;
    cod_agrupador: string;
    agrupador: number;
    cod_variable: string;
    variable: string;
    anio: number;
    anio_uso: number;
}